refactor: split agent k8s config module
